style: use isServiceReady to determine ready status, move Download Logs back outside Overview item, use collapse icon instead of expand, created PageTitle component to be reusable with all views
